% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.est.R
\name{.to.est}
\alias{.to.est}
\title{Return a vector for cm in estimable()}
\usage{
.to.est(obj, params)
}
\arguments{
\item{obj}{estimable object}

\item{params}{character vector of names or logical vector with one element per model parameter selecting desrired parameter(s).}
}
\description{
Return a vector for cm in estimable()
}
\author{
Randy Johnson, Laboratory of Genomic Diversity at NCI-Frederick
}
