% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_p_ci.R
\name{est_p_ci}
\alias{est_p_ci}
\title{Display Confidence interval and p-value for one model term}
\usage{
est_p_ci(model, term, mult = 1, digits = 2, ...)
}
\arguments{
\item{model}{model object}

\item{term}{model term}

\item{mult}{scale (multiply) the parameter by this factor}

\item{digits}{number of significant digits to display}

\item{...}{optional arguments}
}
\description{
Display Confidence interval and p-value for one model term
}
